Guess The Number
Guess The Number is a simple android gaming application which is used to guess the number between the  the range. In this application user will have three attempts. If the user will choose the generated number within three attempt then it will display the toast message saying that "you are genius". If the user fails to guess the number in the particular range within this attempts then  it will display the toast message saying that "reguess the number".